// hub/src/payload-types.ts
export interface BibleChapter {
  id: number;
  book: number | BibleBook;
  name: string;
  order: number;
  globalOrder: number;
  shortName: string;
  slug: string;
  hasNoChapters?: boolean | null;
  verses: number;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "cover-images".
 */
export interface CoverImage {
  id: number;
  name?: string | null;
  alt: string;
  purpose?: ('hub-image' | 'series-image' | 'other')[] | null;
  version?: number | null;
  guid?: string | null;
  sha1sum?: string | null;
  hashInvalid?: boolean | null;
  /**
   * A square SVG of the image to use instead
   */
  squareSvg?: (number | null) | CoverImageSvg;
  /**
   * A 16:9 SVG of the image to use instead
   */
  cardSvg?: (number | null) | CoverImageSvg;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    thumbnail_webp?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    thumbnail_large?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    card?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    card_webp?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    card_large?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    largeSquare?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "cover-image-svgs".
 */
export interface CoverImageSvg {
  id: number;
  alt: string;
  svgFocalPoint?:
    | (
        | 'top-left'
        | 'top-center'
        | 'top-right'
        | 'center-left'
        | 'center-center'
        | 'center-right'
        | 'bottom-left'
        | 'bottom-center'
        | 'bottom-right'
      )
    | null;
  version?: number | null;
  guid?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
}
// ...
export interface TalkAudio {
  id: number;
  originalFilename?: string | null;
  alt?: string | null;
  uploadQuality?: ('best' | 'good' | 'medium' | 'low') | null;
  lengthSeconds?: number | null;
  lengthDisplay?: string | null;
  status?: ('initial' | 'processing' | 'ready') | null;
  qualityGoodMp3_filename?: string | null;
  qualityGoodMp3_mimetype?: string | null;
  qualityGoodMp3_filesize?: number | null;
  qualityGoodOpus_filename?: string | null;
  qualityGoodOpus_mimetype?: string | null;
  qualityGoodOpus_filesize?: number | null;
  qualityMediumMp3_filename?: string | null;
  qualityMediumMp3_mimetype?: string | null;
  qualityMediumMp3_filesize?: number | null;
  qualityMediumOpus_filename?: string | null;
  qualityMediumOpus_mimetype?: string | null;
  qualityMediumOpus_filesize?: number | null;
  qualityLowMp3_filename?: string | null;
  qualityLowMp3_mimetype?: string | null;
  qualityLowMp3_filesize?: number | null;
  qualityLowOpus_filename?: string | null;
  qualityLowOpus_mimetype?: string | null;
  qualityLowOpus_filesize?: number | null;
  guid?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "episodes".
 */
export interface Episode {
  id: number;
  title: string;
  subtitle?: string | null;
  series?: (number | null) | Series;
  episodeType?: ('regular' | 'special') | null;
  biblePassageText?: string | null;
  biblePassages?:
    | {
        chapter: number | BibleChapter;
        verseStart: number;
        verseEnd: number;
        id?: string | null;
      }[]
    | null;
  speaker?: (number | null) | Speaker;
  sermonDate: string;
  sermonDateYear?: number | null;
  /**
   * Ordering for when multiple episodes are on the same day
   */
  order?: number | null;
  episodeImage?: (number | null) | CoverImage;
  publishedAt?: string | null;
  videoFormat: 'vimeo' | 'youtube' | 'none';
  videoUrl?: string | null;
  audioFormat: 'linked' | 'uploaded' | 'none';
  linkedAudioUrl?: string | null;
  linkedAudioFiletype?: string | null;
  linkedAudioFileSize?: number | null;
  linkedAudioLength?: number | null;
  uploadedAudioFile?: (number | null) | TalkAudio;
  hasValidMedia?: boolean | null;
  talkOutline?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  slug: string;
  fullTitle: string;
  guid?: string | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "series".
 */
export interface Series {
  id: number;
  title: string;
  subtitle?: string | null;
  seriesType?: ('regular' | 'special') | null;
  seriesDate: string;
  seriesImage: number | CoverImage;
  slug: string;
  expandedTitle: string;
  episodes?: {
    docs?: (number | Episode)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  guid?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "speakers".
 */
export interface Speaker {
  id: number;
  name?: string | null;
  guid?: string | null;
  talks?: {
    docs?: (number | Episode)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  updatedAt: string;
  createdAt: string;
}